+++ src/Attributes/DoNotDiscover.php
@@ -4,5 +4,5 @@ namespace Poshtive\Router\Attributes;
 
 use Attribute;
 
-#[Attribute(Attribute::TARGET_CLASS)]
+#[Attribute(Attribute::TARGET_CLASS | Attribute::TARGET_METHOD)]
 class DoNotDiscover implements DiscoveryAttribute {}