+++ src/Attributes/Route.php
@@ -10,7 +10,9 @@ class Route implements DiscoveryAttribute
     public function __construct(
         public ?string $uri = null,
         public array|string|null $method = null,
+        public ?string $name = null,
         public array|string|null $middleware = null,
         public bool $keepOrder = false,
+        public bool $absolute = false,
     ) {}
 }